Add Java task solutions and minor fixes
Adds many Java task sources (pattern programs, array exercises, StudentMarksArray, IncrementOperators, and tasks 45–55) and updates class files. Fixes small bugs and cleanups: corrects Scanner import/usage and averaging in Bin.java; read price with nextDouble in ShoppingBill; fix irregular array sizes; simplify ArithmeticOperators initializers; adjust prompts in LargestNumber; comment duplicate prints in PrintDemo; fix If_Marks indentation; rename/cleanup several files (ternary, primitive types, switch case).

+++ 1. Java/Task 43 - Simple Shopping Bill Calculator/ShoppingBill.java
@@ -5,7 +5,7 @@ class ShoppingBill{
 		Scanner sc = new Scanner(System.in);
 		
 		System.out.print("Enter item Price:");
-		double price =sc.nextInt();
+		double price =sc.nextDouble();
 		
 		System.out.print("Enter quantity:");
 		int quantity =sc.nextInt();
